# docs/integrations/api/openapi.yaml
openapi: 3.0.3
info:
  title: Landslide Monitoring API (v2)
  version: 0.1.0
  description: |
    本文件是 API 契约的机器可读入口（契约优先），与 `docs/integrations/api/*.md` 保持一致。
servers:
  - url: /api/v1
...
security:
  - BearerAuth: []
paths:
...
  /data/state/{deviceId}:
    parameters:
      - $ref: "#/components/parameters/DeviceIdPath"
    get:
      tags: [data]
      summary: Get latest device state (shadow)
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DeviceStateResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /data/series/{deviceId}:
    parameters:
      - $ref: "#/components/parameters/DeviceIdPath"
    get:
      tags: [data]
      summary: Get device time series by sensorKeys
      parameters:
        - name: startTime
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: endTime
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: sensorKeys
          in: query
          required: true
          schema: { type: string, description: Comma-separated keys }
        - name: interval
          in: query
          required: false
          schema: { type: string, enum: [raw, 1m, 5m, 1h, 1d], default: raw }
        - name: timeField
          in: query
          required: false
          schema: { type: string, enum: [received, event], default: received }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DeviceSeriesResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /data/raw/{deviceId}:
    parameters:
      - $ref: "#/components/parameters/DeviceIdPath"
    get:
      tags: [data]
      summary: Get raw telemetry points (debug/export)
      parameters:
        - name: startTime
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: endTime
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: sensorKey
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100000, default: 10000 }
        - name: order
          in: query
          required: false
          schema: { type: string, enum: [asc, desc], default: asc }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DeviceRawResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /data/statistics:
    get:
      tags: [data]
      summary: Get aggregated statistics
      parameters:
        - name: scope
          in: query
          required: true
          schema: { type: string, enum: [device, station] }
        - name: deviceId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: stationId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: sensorKey
          in: query
          required: true
          schema: { type: string }
        - name: startTime
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: endTime
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: interval
          in: query
          required: false
          schema: { type: string, enum: [1h, 1d], default: 1h }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StatisticsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /data/export:
    post:
      tags: [data]
      summary: Export data (async recommended)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          $ref: "#/components/responses/SuccessResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DeviceIdPath:
      name: deviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    SuccessResponse:
      description: ok
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
    ErrorResponse:
      description: error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedResponse:
      description: unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ForbiddenResponse:
      description: forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundResponse:
      description: not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerErrorResponse:
      description: internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    SuccessResponse:
      type: object
      required: [success, code, message, timestamp, traceId]
      properties:
        success: { type: boolean, enum: [true] }
        code: { type: integer, example: 200 }
        message: { type: string, example: ok }
        data: {}
        timestamp:
          type: string
          format: date-time
          example: "2025-12-15T10:00:00Z"
        traceId:
          type: string
          example: "req_01J..."
    ErrorResponse:
      type: object
      required: [success, code, message, timestamp, traceId]
      properties:
        success: { type: boolean, enum: [false] }
        code: { type: integer, example: 400 }
        message: { type: string, example: 参数错误 }
        error:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          example: "2025-12-15T10:00:00Z"
        traceId:
          type: string
          example: "req_01J..."
...
    DeviceStateResponse:
      type: object
      required: [deviceId, updatedAt, state]
      properties:
        deviceId: { type: string, format: uuid }
        updatedAt: { type: string, format: date-time }
        state:
          type: object
          required: [metrics]
          properties:
            metrics:
              type: object
              additionalProperties: true
            meta:
              type: object
              additionalProperties: true
...
    DeviceSeriesResponse:
      type: object
      required: [deviceId, startTime, endTime, interval, series]
      properties:
        deviceId: { type: string, format: uuid }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        interval: { type: string }
        series:
          type: array
          items:
            type: object
            required: [sensorKey, points]
            properties:
              sensorKey: { type: string }
              unit: { type: string }
              dataType: { type: string }
              points:
                type: array
                items:
                  type: object
                  required: [ts, value]
                  properties:
                    ts: { type: string, format: date-time }
                    value: {}
        missing:
          type: array
          items:
            type: object
            required: [sensorKey, reason]
            properties:
              sensorKey: { type: string }
              reason: { type: string }
    DeviceRawResponse:
      type: object
      required: [deviceId, sensorKey, list]
      properties:
        deviceId: { type: string, format: uuid }
        sensorKey: { type: string }
        list:
          type: array
          items:
            type: object
            required: [receivedTs, value]
            properties:
              receivedTs: { type: string, format: date-time }
              eventTs: { type: string, format: date-time, nullable: true }
              seq: { type: integer, nullable: true }
              value: {}
              quality: { type: integer, nullable: true }
    StatisticsResponse:
      type: object
      required: [scope, sensorKey, interval, buckets]
      properties:
        scope: { type: string, enum: [device, station] }
        sensorKey: { type: string }
        interval: { type: string }
        buckets:
          type: array
          items:
            type: object
            required: [ts, count]
            properties:
              ts: { type: string, format: date-time }
              min: { type: number, nullable: true }
              max: { type: number, nullable: true }
              avg: { type: number, nullable: true }
              count: { type: integer, minimum: 0 }
    ExportRequest:
      type: object
      required: [scope, startTime, endTime, sensorKeys, format]
      properties:
        scope: { type: string, enum: [device, station] }
        deviceId: { type: string, format: uuid }
        stationId: { type: string, format: uuid }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        sensorKeys:
          type: array
          items: { type: string }
        format: { type: string, enum: [csv, json] }
...